pyproject.toml
[tool.poetry]
name = "petstore_swagger_bdd"
version = "0.1.0"
description = ""
authors = ["tiMoy <timcacayan@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.7"
behave = "^1.2.6"
selenium = "^3.141.0"
PyYAML = "^5.4.1"
expects = "^0.9.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.14.0"
mypy = "^0.910"
black = "^21.7b0"
pep8-naming = "0.8.2"
flake8 = "3.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
